8.18 Refactoring Resolution Modal.
