+++ src/modules/ResolutionModal/components/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateResolutionForm';
