+++ src/modules/ResolutionsList/index.tsx
@@ -36,7 +36,6 @@ export function ResolutionsList() {
 
   const pagesCount = Math.ceil(data.length / RESOLUTION_TABLE_PAGE_SIZE)
   const visiblePagination = createVisiblePagination(pagesCount);
-
   const initialState = {
     pageCount: pagesCount,
     pageIndex: 0,
@@ -57,69 +56,66 @@ export function ResolutionsList() {
     state: { pageIndex },
   } = useTable<any>({ columns, data, initialState: initialState }, usePagination);
 
-
-  if (fetchStatus === FetchStatus.PENDING) {
-    return (
-      <Loader
-        type="Puff"
-        color="#00BFFF"
-        height={100}
-        width={100}
-        timeout={3000}/>
-    )
-  } else {
-    return (
-      <>
-        <ResolutionTableContainer>
-          <ResolutionTable {...getTableProps()}>
-            <ResolutionTableHead>
-            {headerGroups.map(headerGroup => (
-              <ResolutionTableHeadRow {...headerGroup.getHeaderGroupProps()}>
-                {headerGroup.headers.map(column => (
-                  <ResolutionTableHeaderCell {...column.getHeaderProps()}>
-                    {column.render('Header')}
-                  </ResolutionTableHeaderCell>
+  return (
+    <>
+      {fetchStatus === FetchStatus.PENDING
+        ? <Loader
+          type="Puff"
+          color="#00BFFF"
+          height={100}
+          width={100}
+          timeout={3000}/>
+        : <>
+          <ResolutionTableContainer>
+            <ResolutionTable {...getTableProps()}>
+              <ResolutionTableHead>
+                {headerGroups.map(headerGroup => (
+                  <ResolutionTableHeadRow {...headerGroup.getHeaderGroupProps()}>
+                    {headerGroup.headers.map(column => (
+                      <ResolutionTableHeaderCell {...column.getHeaderProps()}>
+                        {column.render('Header')}
+                      </ResolutionTableHeaderCell>
+                    ))}
+                  </ResolutionTableHeadRow>
                 ))}
-              </ResolutionTableHeadRow>
-            ))}
-            </ResolutionTableHead>
-            <tbody {...getTableBodyProps()}>
-            {page.map((row, i) => {
-              prepareRow(row)
-              return (
-                <ResolutionTableRow {...row.getRowProps()}>
-                  {row.cells.map(cell => {
-                    return <ResolutionTableCell {...cell.getCellProps()}>
-                      {cell.render('Cell')}
-                    </ResolutionTableCell>
-                  })}
-                </ResolutionTableRow>
-              )
-            })}
-            </tbody>
-          </ResolutionTable>
-        </ResolutionTableContainer>
+              </ResolutionTableHead>
+              <tbody {...getTableBodyProps()}>
+              {page.map((row, i) => {
+                prepareRow(row)
+                return (
+                  <ResolutionTableRow {...row.getRowProps()}>
+                    {row.cells.map(cell => {
+                      return <ResolutionTableCell {...cell.getCellProps()}>
+                        {cell.render('Cell')}
+                      </ResolutionTableCell>
+                    })}
+                  </ResolutionTableRow>
+                )
+              })}
+              </tbody>
+            </ResolutionTable>
+          </ResolutionTableContainer>
 
-        <ResolutionTablePaginationButtonsContainer>
-          <button onClick={() => previousPage()} disabled={!canPreviousPage}>
-            <img src={BackArrowSvg} width="7" height="14" alt="icon"/>
-          </button>
-          <div>
-            {visiblePagination
-              .map((paginationNumber) => (
-              <ResolutionTablePaginationButton
-                onClick={() => gotoPage(paginationNumber - 1)}
-                isActive={paginationNumber === pageIndex + 1}
-                key={`pagination-button-${paginationNumber}`}>
-                {paginationNumber}
-              </ResolutionTablePaginationButton>
-            ))}
-          </div>
-          <button onClick={() => nextPage()} disabled={!canNextPage}>
-            <img src={RightArrowSvg} width="7" height="14" alt="icon"/>
-          </button>
-        </ResolutionTablePaginationButtonsContainer>
-      </>
-    )
-  }
+          <ResolutionTablePaginationButtonsContainer>
+            <button onClick={() => previousPage()} disabled={!canPreviousPage}>
+              <img src={BackArrowSvg} width="7" height="14" alt="icon"/>
+            </button>
+            <div>
+              {visiblePagination
+                .map((paginationNumber) => (
+                  <ResolutionTablePaginationButton
+                    onClick={() => gotoPage(paginationNumber - 1)}
+                    isActive={paginationNumber === pageIndex + 1}
+                    key={`pagination-button-${paginationNumber}`}>
+                    {paginationNumber}
+                  </ResolutionTablePaginationButton>
+                ))}
+            </div>
+            <button onClick={() => nextPage()} disabled={!canNextPage}>
+              <img src={RightArrowSvg} width="7" height="14" alt="icon"/>
+            </button>
+          </ResolutionTablePaginationButtonsContainer>
+        </>}
+    </>
+  )
 }
